requirements.txt
flask-socketio
python-dotenv